+++ setup.cfg
@@ -54,10 +54,10 @@ per-file-ignores =
   # We allow reexport:
   returns/pointfree/__init__.py: F401, WPS201
   returns/methods/__init__.py: F401, WPS201
-  returns/pipeline.py: F401
   returns/context/__init__.py: F401, WPS201
+  returns/io/__init__.py: F401
+  returns/pipeline.py: F401
   # Disable some quality checks for the most heavy parts:
-  returns/io.py: WPS402
   returns/iterables.py: WPS234
   # Interfaces and asserts can have assert statements:
   returns/interfaces/*.py: S101
